backend/requirements.txt
Django == 6.0
djangorestframework == 3.16.1
djangorestframework-jwt == 1.11.0
django-cors-headers == 4.9.0
openpyxl == 3.1.5